rename previousStep to previous-step
